requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0
requests>=2.31.0
